backend_fastapi/requirements.txt
absl-py==0.12.0
aiofiles==0.6.0
asgiref==3.3.1
astunparse==1.6.3
blis==0.7.4
cachetools==4.2.1
catalogue==2.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
comtypes==1.1.8
cryptography==3.3.1
cymem==2.0.5
Django==3.1.7
djangorestframework==3.12.4
docx2txt==0.8
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.0.0/en_core_web_lg-3.0.0-py3-none-any.whl
et-xmlfile==1.0.1
fastapi==0.63.0
flatbuffers==1.12
gast==0.4.0
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
gTTS==2.2.2
gunicorn==20.1.0
h11==0.12.0
h2==4.0.0
h5py==3.1.0
hpack==4.0.0
Hypercorn==0.11.2
hyperframe==6.0.0
idna==2.10
jdcal==1.4.1
Jinja2==2.11.3
joblib==1.0.1
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==1.1.1
murmurhash==1.0.5
nltk==3.5
numpy==1.19.5
oauthlib==3.1.0
openpyxl==3.0.5
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.5
pathy==0.4.0
pdfminer.six==20201018
playsound==1.2.2
preshed==3.0.5
priority==1.3.0
protobuf==3.15.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.3
PyMySQL==0.10.1
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
python-dotenv==0.17.0
python-multipart==0.0.5
pyttsx3==2.90
pytz==2020.4
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.2
six==1.15.0
smart-open==3.0.0
sortedcontainers==2.3.0
spacy==3.0.3
spacy-legacy==3.0.1
SQLAlchemy==1.3.20
sqlparse==0.4.1
srsly==2.4.0
starlette==0.13.6
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0rc0
termcolor==1.1.0
tf-estimator-nightly==2.5.0.dev2021032501
thinc==8.0.1
toml==0.10.2
tqdm==4.56.2
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn==0.13.4
wasabi==0.8.2
watchgod==0.7
websockets==8.1
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
wsproto==1.0.0
xlrd==1.2.0
